+++ src/atoms/types.d.ts
@@ -4,6 +4,11 @@ export interface Choice {
   value: string,
 }
 
+export interface Option {
+  label: string,
+  value: string,
+}
+
 export type PositionType = "top-left" | "bottom-left" | "top-right" | "bottom-right";
 
 export type IframeSandbox = "allow-forms" | "allow-modals" | "allow-pointer-lock" | "allow-popups" | "allow-popups-to-escape-sandbox" | "allow-same-origin" | "allow-scripts" | "allow-top-navigation";