+++ CHANGELOG.md
@@ -9,6 +9,7 @@ title: Changelog
 * `<cc-tcp-redirection-form>`: add smart definition
 * New component:
   * `<cc-select>`
+* `<cc-button>`, `<cc-input-number>`, `<cc-input-text>`, `<cc-select>`, `<cc-toggle>`: add new story showing all form controls together
 
 ## 7.8.0 (2022-03-10)