+++ CHANGELOG.md
@@ -14,12 +14,16 @@ title: Changelog
 * Remove `engines.node` from `package.json` and rely on Volta's versions fields
   * We use the `CC_PRE_BUILD_HOOK` env var to select the versions for Node.js and npm
 
+### For devs
+
+* Use specific Docker image for the CDN deployment job
+
 ## 7.8.0 (2022-03-10)
 
 * `<cc-env-var-*>`: fix links to the doc
 * `<cc-pricing-product>`: add details for Heptapod runners
 
-## For devs
+### For devs
 
 * Improve the CEM analyzer generate behaviour in dev mode