+++ CHANGELOG.md
@@ -17,6 +17,7 @@ title: Changelog
 ### For devs
 
 * Use specific Docker image for the CDN deployment job
+* Fix npm task components:check-lit (was only applied on atoms)
 
 ## 7.8.0 (2022-03-10)