+++ CHANGELOG.md
@@ -13,6 +13,7 @@ title: Changelog
 * `<cc-invoice-list>`: implement `<cc-select>` to display the year list when component width is lower than 520px
 * Remove `engines.node` from `package.json` and rely on Volta's versions fields
   * We use the `CC_PRE_BUILD_HOOK` env var to select the versions for Node.js and npm
+* `<cc-header-addon>`: add missing import to cc-zone
 
 ### For devs